requirements.txt
asgiref==3.3.1
autopep8==1.5.5
boto3==1.17.17
botocore==1.20.17
Django==3.1.6
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.1.0
pycodestyle==2.6.0
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
